Reorganize core common crate
