+++ nautilus_core/common/src/python/mod.rs
@@ -12,3 +12,5 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
+
+pub mod timer;
