+++ nautilus_trader/core/rust/common.pxd
@@ -113,24 +113,6 @@ cdef extern from "../includes/common.h":
     cdef struct TestClock:
         pass
 
-    # Represents a time event occurring at the event timestamp.
-    cdef struct TimeEvent_t:
-        # The event name.
-        char* name;
-        # The event ID.
-        UUID4_t event_id;
-        # The message category
-        uint64_t ts_event;
-        # The UNIX timestamp (nanoseconds) when the object was initialized.
-        uint64_t ts_init;
-
-    # Represents a time event and its associated handler.
-    cdef struct TimeEventHandler_t:
-        # The event.
-        TimeEvent_t event;
-        # The event ID.
-        PyObject *callback_ptr;
-
     # Provides a C compatible Foreign Function Interface (FFI) for an underlying [`TestClock`].
     #
     # This struct wraps `TestClock` in a way that makes it compatible with C function
@@ -165,6 +147,24 @@ cdef extern from "../includes/common.h":
     cdef struct Logger_API:
         Logger_t *_0;
 
+    # Represents a time event occurring at the event timestamp.
+    cdef struct TimeEvent_t:
+        # The event name.
+        char* name;
+        # The event ID.
+        UUID4_t event_id;
+        # The message category
+        uint64_t ts_event;
+        # The UNIX timestamp (nanoseconds) when the object was initialized.
+        uint64_t ts_init;
+
+    # Represents a time event and its associated handler.
+    cdef struct TimeEventHandler_t:
+        # The event.
+        TimeEvent_t event;
+        # The event ID.
+        PyObject *callback_ptr;
+
     const char *component_state_to_cstr(ComponentState value);
 
     # Returns an enum from a Python string.
@@ -196,8 +196,6 @@ cdef extern from "../includes/common.h":
     # # Safety
     # - Assumes `ptr` is a valid C string pointer.
     LogColor log_color_from_cstr(const char *ptr);
-
-    TimeEventHandler_t dummy(TimeEventHandler_t v);
 
     TestClock_API test_clock_new();
 
@@ -324,3 +322,5 @@ cdef extern from "../includes/common.h":
 
     # Returns a [`TimeEvent`] as a C string pointer.
     const char *time_event_to_cstr(const TimeEvent_t *event);
+
+    TimeEventHandler_t dummy(TimeEventHandler_t v);
